docs: split up storybook stories
